Finished CRUD basic (not include dto)

+++ BE/oem-ev-warranty-management-system/src/main/java/com/swp391/warrantymanagement/entity/WarrantyClaims.java
@@ -17,7 +17,7 @@ public class WarrantyClaims {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "warranty_claim_id")
-    private Long warrantyClaimId;
+    private int warrantyClaimId;
 
     @Column(name = "claim_date", nullable = false)
     private Date claimDate;
